+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/NullBuilder.java
@@ -35,7 +35,7 @@ final class NullBuilder implements Builder {
   }
 
   @Override
-  public int getCurrentSize() {
+  public long getCurrentSize() {
     return length;
   }
