+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/TypedBuilder.java
@@ -75,7 +75,7 @@ abstract class TypedBuilder<T> implements BuilderWithRetyping, BuilderForType<T>
   }
 
   @Override
-  public int getCurrentSize() {
+  public long getCurrentSize() {
     return currentSize;
   }
